requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-filters==0.9.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.8
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.4
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
